+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './litellm-api'
 export * from './format-model-name'
+export * from './model-filter'